requirements.txt
arxiv
ollama
python-dateutil
requests
PyPDF2
PyYAML
